+++ src/ingest/index.js
@@ -1,7 +1,7 @@
 /*
 
 	ard-eventhub
-	by SWR audio lab
+	by SWR Audio Lab
 
 */
 
@@ -26,17 +26,19 @@ server.use((req, res, next) => {
 	// add service information
 	res.set('x-service', config.userAgent)
 
-	// DEV temporarily log all headers to validate ingress
-	const logHeaders = {
-		...req.headers,
-		authorization: 'hidden',
+	// log all headers in local mode
+	if (config.isLocal) {
+		const logHeaders = {
+			...req.headers,
+			authorization: 'hidden',
+		}
+		logger.log({
+			level: 'debug',
+			message: `middleware logging`,
+			source: 'DEV',
+			data: { logHeaders, path: req.path },
+		})
 	}
-	logger.log({
-		level: 'debug',
-		message: `middleware logging`,
-		source: 'DEV',
-		data: logHeaders,
-	})
 
 	// continue with normal workflow
 	next()
@@ -50,7 +52,7 @@ server.use(compression())
 server.disable('x-powered-by')
 server.listen(config.port)
 
-if (config.isDev) {
+if (config.isLocal) {
 	console.log(`${config.serviceName} (v${config.version}) is running at: ${config.serviceUrl}`)
 	console.log(`  - OpenAPI documentation: ${config.serviceUrl}/openapi`)
 }