+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - fix: stop logging requests on dev
 - fix: spelling of `SWR Audio Lab`
+- fix: handle invalid format definition
 - feat: add `IS_LOCAL` env for local-specific logs and actions
 - feat: enable multiple user accounts for DTS plugin
 - feat: add test for invalid timestamps