+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.6.0] - 2023-09-22
 
+- feat: make dts plugin opt-out
 - refactor: migrate DTS APIs
 
 ## [1.5.2] - 2023-08-08