+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.8.0] - 2024-06-26
+
+- chore: upgrade to Node.js 20
+- chore: swap yarn for bun (package manager only)
+- chore: remove table of contents from docs
+
 ## [1.7.3] - 2024-04-23
 
 - chore: updated lots of dependencies